ECA-7573: Fix review comments. Remove more unused code
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@30856 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/admin-gui/src/org/ejbca/ui/web/admin/AdminIndexMBean.java
@@ -80,7 +80,7 @@ public class AdminIndexMBean extends BaseManagedBean implements Serializable {
         final List<CaCrlStatusInfo> ret = new ArrayList<>();
         final Collection<Integer> caIds = caSession.getAuthorizedCaIds(getAdmin());
         for (final Integer caId : caIds) {
-            final CAInfo cainfo = caSession.getCAInfoInternal(caId.intValue());
+            final CAInfo cainfo = caSession.getCAInfoInternal(caId);
             if (cainfo == null || cainfo.getStatus() == CAConstants.CA_EXTERNAL) {
                 continue;
             }
